Standardize Economy Names and Codes

Standardizes economy names and codes in a dataset by matching them against a predefined list of patterns and ISO codes. Handles aggregate economies.

Usage

standardize_economies(
  data,
  name_col,
  code_col = NULL,
  output_cols = c("economy_id", "economy_name", "economy_type"),
  default_economy_type = NA_character_,
  warn_ambiguous = TRUE,
  overwrite = TRUE,
  warn_overwrite = TRUE
)

Arguments

data

A data frame or tibble containing economy names to standardize

name_col

Name of the column containing economy names

code_col

Optional name of the column containing economy codes

output_cols

Character vector specifying desired output columns. Options are "economy_name", "economy_id", "economy_type", "iso3c", "iso2c"

default_economy_type

Character; the default economy type to use if not specified in the data. Options are "country", "institution", or "aggregate". Defaults to NA. Will be ignored if output_cols do not include "economy_type".

warn_ambiguous

Logical; whether to warn about ambiguous matches

overwrite

Logical; whether to overwrite existing economy_* columns

warn_overwrite

Logical; whether to warn when overwriting existing economy_* columns. Defaults to TRUE.

Value

A data frame with standardized economy information merged with the input data

Examples

if (FALSE) { # \dontrun{
df <- data.frame(economy = c("United States", "China"))
standardize_economy(df, name_col = economy)
} # }