Standardize entity Names and Codes

Standardizes entity names and codes in a dataset by matching them against a predefined list of patterns and ISO codes. Handles aggregate entities.

Usage

standardize_entities(
  data,
  name_col,
  code_col = NULL,
  output_cols = c("entity_id", "entity_name", "entity_type"),
  default_entity_type = NA_character_,
  warn_ambiguous = TRUE,
  overwrite = TRUE,
  warn_overwrite = TRUE
)

Arguments

data

A data frame or tibble containing entity names to standardize

name_col

Name of the column containing entity names

code_col

Optional name of the column containing entity codes

output_cols

Character vector specifying desired output columns. Options are "entity_name", "entity_id", "entity_type", "iso3c", "iso2c". Defaults to c("entity_id", "entity_name", "entity_type").

default_entity_type

Character; the default entity type to use for entities that do not match any of the patterns. Options are "economy", "organization", "aggregate", or "other". If this argument is not supplied, the default value will be NA_character_. Argument will be ignored if output_cols do not include "entity_type".

warn_ambiguous

Logical; whether to warn about ambiguous matches

overwrite

Logical; whether to overwrite existing entity_* columns

warn_overwrite

Logical; whether to warn when overwriting existing entity_* columns. Defaults to TRUE.

Value

A data frame with standardized entity information merged with the input data

Examples

if (FALSE) { # \dontrun{
df <- data.frame(entity = c("United States", "China"))
standardize_entity(df, name_col = entity)
} # }